Add files via upload
From the Pyber Ride Sharing Data (2016) it is clearly seen in the bubble plot that for the fact that there are less drivers in rural cities, drivers in such cities still make more income than the suburban and urban cities drivers. 
In the fares by city pie chart, it also seen that more than half of the fares charged are been paid in the urban cities compared to the suburban and rural cities.
In the rides by city type pie chart, clear majority of rides are from the urban cities compared to the suburban cities which come second in rides and the rural cities comes last.
In the total drivers by city type, it is also clearly seen that more drivers are in the urban cities compared to the suburban cities and the rural cities.
In conclusion drivers who have the preferences of driving more should drive in both the urban and suburban cities because there are available riders in such cities, and drivers who want to the highest income should drive in the rural cities.